Removed `git rev-parse` call
It was checking whether we were in a git repo.  However, the open() call
to open `git log` also tells us that, so the rev-parse check was
redundant.  The only other `git` call before the `git log` is a
`git --exec-path`, which doesn't care whether or not we're in a repo.

I also added some comments to mark major parts of the script.